+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ================
 * [#1098](https://github.com/oshi/oshi/pull/1098): Option to limit FileStore list to local file systems. - [@Space2Man](https://github.com/Space2Man).
 * [#1100](https://github.com/oshi/oshi/pull/1100): Get FileStore options. - [@dbwiddis](https://github.com/dbwiddis).
+* [#1101](https://github.com/oshi/oshi/pull/1101): Add network interface dropped packets and collisions. - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here.
 
 4.3.0 (1/2/2020), 4.3.1 (2/5/2020)